refactor: #102 - signup mutation 응답 계약 보존 및 구조 분리
